combined collections all & collections handle UI into a template component
